partner/pyproject.toml
[tool.poetry]
name = "amqp-grpc"
version = "0.1.0"
description = ""
authors = [""]
readme = "README.md"
packages = [{include = "amqp_grpc"}]

[tool.poetry.dependencies]
python = "^3.10"
boto3 = "^1.28.28"
pika = "^1.3.2"
pytest = "^7.4.0"
grpcio = "^1.57.0"
google = "^3.0.0"
protobuf = "^4.24.0"
psycopg = "^3.1.10"
PyJWT = "^2.8.0"
requests = "^2.31.0"
cryptography = "^41.0.3"
grpc-interceptor = "^0.15.3"
cognitojwt = "^1.4.1"
epyestim = "^0.1"
numpy = "^1.25.2"
pandas = "^2.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
